edit

+++ src/main/java/com/action/CrudAction.java
@@ -1,5 +1,7 @@
 package com.action;
 
+import java.lang.reflect.Field;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -21,9 +23,13 @@ import com.common.page.Page;
 import com.common.page.PageDataTrans;
 import com.common.page.PageUtil;
 import com.common.response.ResponseModel;
+import com.common.spring.BeanHelper;
 import com.common.utils.Const;
+import com.common.utils.LogUtil;
+import com.common.utils.TableUtil;
 import com.model.BaseModel;
 import com.service.BaseService;
+import com.service.system.TableService;
 
 
 /**  
@@ -38,6 +44,57 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 	
 	@Autowired
 	protected PageUtil pageUtil;
+	
+	/**
+	 * 对列表表格样式进行处理
+	 * 2019年2月12日
+	 * @author:Lynn
+	 */
+	public void handleList(Page page){
+		//每列宽度
+		Map<Integer,String> widthMap = TableUtil.getWidthMap(model.getClass().getName());
+		pageUtil.setColsMinWidthMap(widthMap);
+		//tpl模板
+		Map<Integer,String> tplMap = TableUtil.getTplMap(model.getClass().getName());
+		pageUtil.setColsTempletMap(tplMap);
+		//style样式
+		Map<Integer,String> styleMap = TableUtil.getStyleMap(model.getClass().getName());
+		pageUtil.setColsStyleMap(styleMap);
+		//设置列是否隐藏
+		Map<Integer,Boolean> hideMap = TableUtil.getHideMap(model.getClass().getName());
+		pageUtil.setColsHideMap(hideMap);
+		//设置列是否可修改
+		Map<Integer,String> inputMap = TableUtil.getInputMap(model.getClass().getName());
+		pageUtil.setColsEditMap(inputMap);
+		//设置操作列的宽度
+		if(StringUtils.isNotEmpty(widthMap.get(0)) && !widthMap.get(0).equals("0")){
+			pageUtil.setColsWidth(page.getHeader().length, widthMap.get(0));
+		}
+		
+	}
+	
+	/**
+	 * 对列表分页数据进行处理
+	 * 2019年2月12日
+	 * @author:Lynn
+	 */
+	public void handleListData(){
+		//数字字典
+		Map<Integer,String> dictMap = TableUtil.getDictMap(model.getClass().getName());
+		//tpl模板
+		Map<Integer,String> tplMap = TableUtil.getTplMap(model.getClass().getName());
+		
+		Set<Integer> dictSet = new HashSet<Integer>(dictMap.keySet());
+		//剔除开关控制和tpl模板数据字典的应用
+		for(Integer dictKey:dictSet){
+			if(tplMap.containsKey(dictKey)) dictMap.remove(dictKey);
+		}
+		pageUtil.setDataDictMap(dictMap);
+		
+		//js方法
+		Map<Integer,String> jsfuncMap = TableUtil.getJsfuncMap(model.getClass().getName());
+		pageUtil.setColsJsfuncMap(jsfuncMap);
+	}
 	
 	/**
 	 * 该方法用于处理页面的首页请求，本方法提供一个缺省实现
@@ -62,7 +119,7 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 	@SuppressWarnings({ "rawtypes", "unchecked" })
 	@RequestMapping("/list")
 	public ModelAndView list(){
-		ModelAndView mv = this.getModelAndView();
+		ModelAndView mv = this.getModelAndView(Const.LIST_QUERY_PARAMS);
 		init(mv);
 		//获取列表页面显示数据sql/hql语句
 		Map sqlMap = (Map)parasMap.get(Const.SQL_MAP);
@@ -86,10 +143,13 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 			logger.error("获取分页数据出错，出错语句："+pageSql, e);
 		}
 		//转换列表数据
-		handleListData();
+		TableUtil.clearTableCache();
+		handleList(page);
 		page = PageDataTrans.TransPageCols(page, pageUtil);
 		mv.addObject(page);
 		service.getUserAuth(getSessionUser(request).getUserId(),(Set<String>) request.getSession().getAttribute(Const.SESSION_AUTH),mv);
+		//获取查询区域
+		mv.addObject(Const.LIST_QUERY,TableUtil.getSearchArea(model.getClass().getName()));
 		return mv;
 	}
 	
@@ -155,7 +215,12 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 	public ModelAndView detail(@RequestParam(value="id", required=false, defaultValue="") String id){
 		//获取到主键值，则加载对象
 		if(StringUtils.isNotEmpty(id)){
-			 model = (M) service.loadModel(model.getClass(), id);
+			 Class<?> modelType = service.getPkType(model.getClass());
+			 if(modelType == Integer.class){
+				 model = (M) service.loadModel(model.getClass(), Integer.parseInt(id));
+			 }else{
+				 model = (M) service.loadModel(model.getClass(), id);
+			 }
 		}
 		ModelAndView mv = this.getModelAndView(model);
 		return mv; 
@@ -173,11 +238,13 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
     @ResponseBody
 	public ResponseModel save(M m){
 		try {
+			LogUtil.remarkAddLog(request, m);
 			service.beforeSaveModel(m);
 			service.saveModel(m);
 			service.afterSaveModel(m);
 			return new ResponseModel(this.getModelAndView(m)); 
 		}catch (Exception e) {
+			LogUtil.setFailLog(request);
 			e.printStackTrace();
 			logger.error("保存"+m.getClass().getName()+"对象出错！");
 			return new ResponseModel(Const.CRUD_ERROR,e.getMessage());
@@ -196,13 +263,18 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 	@ResponseBody
 	public ResponseModel update(M m){
 		try {
-			service.beforeUpdateModel(m);
+			logger.info(m);
+			LogUtil.remarkUpdateLog(request, m);
+//			service.beforeUpdateModel(m);
 			service.updateModel(m);
-			service.afterUpdateModel(m);
+//			service.afterUpdateModel(m);
 			ModelAndView mv = this.getModelAndView();
 			mv.clear();
-			return new ResponseModel(this.getModelAndView(m)); 
+			mv = this.getModelAndView(m);
+			return new ResponseModel();
+//			return new ResponseModel(this.getModelAndView(m)); 
 		}catch (Exception e) {
+			LogUtil.setFailLog(request);
 			e.printStackTrace();
 			logger.error("修改"+m.getClass().getName()+"对象出错！");
 			return new ResponseModel(Const.CRUD_ERROR,e.getMessage());
@@ -221,13 +293,30 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
     @ResponseBody
 	public ResponseModel delete(@PathVariable(value="id") String id){
 		try {
-			if(service.beforeDeleteModel(id)){
-				service.deleteModelById(model.getClass(), id);
-				service.afterDeleteModel(id);
-				return new ResponseModel();
+			if(StringUtils.isNotEmpty(id)){
+				Class<?> modelType = service.getPkType(model.getClass());
+				if(modelType == Integer.class){
+					int pk = Integer.parseInt(id);
+					if(service.beforeDeleteModel(pk)){
+						LogUtil.remarkDelLog(request, model.getClass(), pk);
+						service.deleteModelById(model.getClass(), pk);
+						service.afterDeleteModel(pk);
+						return new ResponseModel();
+					}
+				}else{
+					if(service.beforeDeleteModel(id)){
+						LogUtil.remarkDelLog(request, model.getClass(), id);
+						service.deleteModelById(model.getClass(), id);
+						service.afterDeleteModel(id);
+						return new ResponseModel();
+					}
+				}
+				return new ResponseModel(Const.CRUD_ERROR,"删除失败");
+			}else{
+				return new ResponseModel(Const.PARA_ERROR,Const.FAIL_NO_VALUE);
 			}
-			return new ResponseModel(Const.CRUD_ERROR,"删除失败");
 		} catch (Exception e) {
+			LogUtil.setFailLog(request);
 			e.printStackTrace();
 			logger.error("删除"+model.getClass().getName()+"对象出错！");
 			return new ResponseModel(Const.CRUD_ERROR,e.getMessage());
@@ -248,14 +337,45 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 		try {
 			ModelAndView mv = this.getModelAndView(model);
 			String flag = mv.getModel().get("flag").toString();
+			String column = mv.getModel().get("field").toString();
 			//获取到主键值，则加载对象
 			if(StringUtils.isNotEmpty(id)){
-				 model = (M) service.loadModel(model.getClass(), id);
-				 if("true".equals(flag)){
-					 model.setDataFlag("Y");
+				 Class<?> modelType = service.getPkType(model.getClass());
+				 if(modelType == Integer.class){
+					 model = (M) service.loadModel(model.getClass(), Integer.parseInt(id));
 				 }else{
-					 model.setDataFlag(null);
+					 model = (M) service.loadModel(model.getClass(), id);
 				 }
+				 //获取字段对象
+				 TableService tableService = BeanHelper.getBean("tableManager");
+				 Field field = tableService.getTableField(model.getClass(), column);
+				 if(field == null)  return new ResponseModel(Const.PARA_ERROR,Const.FAIL_EDIT_VALUE);
+				 field.setAccessible(true);
+				 //字段类型比较，设置新的字段值
+				 Class type = field.getType();
+				 if(type == Integer.class || type == int.class || type == Long.class){
+					 if("true".equals(flag)){
+						 field.set(model, 1);
+					 }else{
+						 field.set(model, 0);
+					 }
+				 }
+				 else if(type == Byte.class){
+					 if("true".equals(flag)){
+						 field.set(model, Byte.valueOf("1"));
+					 }else{
+						 field.set(model, Byte.valueOf("0"));
+					 }
+				 }
+				 else{//string类型
+					 
+					 if("true".equals(flag)){
+						 field.set(model, "Y");
+					 }else{
+						 field.set(model, "N");
+					 }
+				 }
+				 field.setAccessible(false);
 				 model.setModifyId(getSessionUser(request).getUserId());
 				 service.updateModel(model);
 				 return new ResponseModel(); 
@@ -263,6 +383,62 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 				return new ResponseModel(Const.PARA_ERROR,Const.FAIL_NO_VALUE);
 			}
 		}catch(Exception e) {
+			LogUtil.setFailLog(request);
+			e.printStackTrace();
+			logger.error("设置数据状态时出错！");
+			return new ResponseModel(Const.CRUD_ERROR,e.getMessage());
+		}
+		
+	}
+	
+	/**
+	 * 该方法用于设置数据状态dataFlag的启用/禁用状态
+	 * 2017年9月22日
+	 * @param id
+	 * @return
+	 * @author:wangzhen
+	 */
+	@SuppressWarnings("all")
+	@RequestMapping(value="/editField",method=RequestMethod.POST)
+	@ResponseBody
+	public ResponseModel editField(@RequestParam(value="id", required=false, defaultValue="") String id,@RequestParam(value="data", required=false, defaultValue="") String data,@RequestParam(value="field", required=false, defaultValue="") String column){
+		try {
+			//获取到主键值，则加载对象
+			if(StringUtils.isNotEmpty(id)){
+				 Class<?> modelType = service.getPkType(model.getClass());
+				 if(modelType == Integer.class){
+					 model = (M) service.loadModel(model.getClass(), Integer.parseInt(id));
+				 }else{
+					 model = (M) service.loadModel(model.getClass(), id);
+				 }
+				 //获取字段对象
+				 TableService tableService = BeanHelper.getBean("tableManager");
+				 Field field = tableService.getTableField(model.getClass(), column);
+				 if(field == null)  return new ResponseModel(Const.PARA_ERROR,Const.FAIL_EDIT_VALUE);
+				 field.setAccessible(true);
+				 //字段类型比较，设置新的字段值
+				 Class type = field.getType();
+				 if(type == Integer.class || type == int.class){
+					 field.set(model, Integer.parseInt(data));
+				 }
+				 else if(type == Long.class){
+					 field.setLong(model, Long.valueOf(data));
+				 }
+				 else if(type == Byte.class){
+					 field.setByte(model, Byte.valueOf(data));
+				 }
+				 else{//string类型
+					 field.set(model, data);
+				 }
+				 field.setAccessible(false);
+				 model.setModifyId(getSessionUser(request).getUserId());
+				 service.updateModel(model);
+				 return new ResponseModel(); 
+			}else{
+				return new ResponseModel(Const.PARA_ERROR,Const.FAIL_NO_VALUE);
+			}
+		}catch(Exception e) {
+			LogUtil.setFailLog(request);
 			e.printStackTrace();
 			logger.error("设置数据状态时出错！");
 			return new ResponseModel(Const.CRUD_ERROR,e.getMessage());
@@ -311,7 +487,6 @@ public class CrudAction<T extends BaseService,M extends BaseModel> extends BaseA
 		mv.addObject("abb", "abbssdsds");
 		request.setAttribute("bbb", "sss");
 		mv.setViewName("forward:"+request.getRequestURI().replace("tree", "list"));
-		System.out.println(mv.getModel().get("abb"));
 		return mv;
 	}
